Fix for https://github.com/lapcat/Bonjeff/issues/3
Make the window contentView layer-backed.

This is a workaround for what appears to be an AppKit bug with NSOutlineView. The outline view draws square corners at the bottom of the window. This issue does not occur if the outline view bottomAnchor is constrained some constant higher than the contentView bottomAnchor.